fass_gateway/requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/
--trusted-host mirrors.aliyun.com

fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.0
httpx==0.28.1
sse-starlette==2.1.3
sentence-transformers==3.3.1
cryptography==42.0.8
